+++ src/test/java/roomescape/reservation/repository/H2ReservationRepositoryTest.java
@@ -12,6 +12,8 @@ import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
 import roomescape.reservation.model.Reservation;
 import roomescape.reservation.model.ReservationTime;
+import roomescape.reservation.repository.h2.H2ReservationRepository;
+import roomescape.reservation.repository.h2.H2ReservationTimeRepository;
 
 @JdbcTest(properties = "application-test.properties")
 @Import({H2ReservationRepository.class, H2ReservationTimeRepository.class})
